requirements/base.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
psycopg2-binary==2.9.9
channels==4.1.0
channels-redis==4.2.0
Pillow==10.4.0
celery==5.4.0
redis==5.0.8
django-treebeard==4.7.1
django-guardian==2.4.0